fix: Added format menu on top right of rich modal
